server/requirements.txt
attrs==21.2.0
python-dotenv==0.19.0
websockets==10.0
